# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TetraCore Hub API
  description: |
    TetraCore Hub is a distributed task orchestration and management system.

    ## Authentication
    The API supports two authentication methods:
    - **JWT Token**: Use `Authorization: Bearer <token>` header
    - **API Key**: Use `X-API-Key: <key>` header or `Authorization: Bearer tk_<key>`

    ## Rate Limiting
    Default rate limit is 1000 requests per minute. API keys may have custom limits.
  version: 2.0.0
  contact:
    name: TetraCore Support
    email: support@tetra-core.website
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://hub.tetra-core.website
    description: Production
...
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
...
  /api/v1/clients/{id}:
    get:
      tags: [Clients]
      summary: Get client
      operationId: getClient
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Client details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
...
  /api/v1/apikeys:
    get:
      tags: [API Keys]
      summary: List API keys
      operationId: listAPIKeys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: API key list
    post:
      tags: [API Keys]
      summary: Create API key
      operationId: createAPIKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAPIKeyRequest'
      responses:
        '201':
          description: API key created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    Client:
      type: object
      properties:
        client_id:
          type: string
        status:
          type: string
        connected_at:
          type: string
          format: date-time
...
    CreateAPIKeyRequest:
      type: object
      required: [name, role_id]
      properties:
        name:
          type: string
        role_id:
          type: string
        expires_in:
          type: string
...